requirements.txt
accelerate==0.27.2
bitsandbytes==0.41.2.post2
black==23.1.0
datasets==2.14.6
deepspeed==0.12.2
einops>=0.6.1
evaluate==0.4.0
flake8>=6.0.0
hf-doc-builder>=0.4.0
hf_transfer>=0.1.4
huggingface-hub>=0.19.2,<1.0
isort>=5.12.0
ninja>=1.11.1
numpy>=1.24.2
packaging>=23.0
parameterized>=0.9.0
peft==0.7.1
protobuf<=3.20.2
pytest
safetensors>=0.3.3
sentencepiece>=0.1.99
scipy
tensorboard
torch==2.1.2
transformers==4.41.1
trl==0.8.6
jinja2>=3.0.0
tqdm>=4.64.1
install
einops
ftfy
h5py
imageio
imageio-ffmpeg
regex
scikit-video
prettytable
deepspeed-mii==0.2.3
pytorch-lightning>=2.0
